fix(scraping): salvage sparse guessed-market hits, probe apify slot, split aggregator module
Review follow-ups: sub-floor guessed-market adzuna hits are returned
when the fallback is unconfigured or fails (empty results still surface
the diagnostic error); the store-error probe now covers the apify token
slot so a faulting slot cannot masquerade as a needs-keys skip. Also
splits the provider implementations into aggregator/providers.rs to
satisfy the architecture loc cap (mod.rs 1516 -> 569).
